Move all iOS-specific code to a module
