+++ Fruitties/shared/src/commonMain/kotlin/com/example/fruitties/model/Fruittie.kt
